requirements.txt
aiohttp
aiofiles
discord.py
numpy
requests
pillow
iso3166
pypinyin
opencv-python
iso3166
